% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_functions.R
\name{add_set_intersection_to_df}
\alias{add_set_intersection_to_df}
\title{Create a vector containing the weighted percentages in each set}
\usage{
add_set_intersection_to_df(data, varnames, exclude_unique = T)
}
\arguments{
\item{data}{a dataframe containing all the sets for each record, coercible to 1,0}

\item{varnames}{a vector containing the names of variables to be used in the intersection}

\item{exclude_unique}{whether the set intersections should include singular sets (i.e. that one variable). Note that if this is set to True, the total set size on the left will be wrong}
}
\value{
A vector of the aggregated percent for each intersection
}
\description{
Create a vector containing the weighted percentages in each set
}
