% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_functions.R
\name{expand_to_set_intersections}
\alias{expand_to_set_intersections}
\title{Expand the binary indicators to set intersections}
\usage{
expand_to_set_intersections(data, varnames)
}
\arguments{
\item{data}{a dataframe containing all the 1,0 indicators in varnames}

\item{varnames}{a vector containing the names of variables to be used in the intersection}
}
\value{
The dataframe containing the intersections of all variables and the names of those combinations of varnames with '&'
}
\description{
Expand the binary indicators to set intersections
}
