% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_functions.R
\name{make_set_percentages}
\alias{make_set_percentages}
\title{Create a vector containing the weighted percentages in each set}
\usage{
make_set_percentages(data, varnames, exclude_unique = T)
}
\arguments{
\item{data}{a dataframe containing all the sets for each record, coercible to 1,0}

\item{nsets}{number of sets to look at}

\item{nsets}{number of intersections to look at, the default being 12}

\item{label}{the label to be added to the plot}
}
\value{
A vector of the aggregated percent for each intersection
}
\description{
Create a vector containing the weighted percentages in each set
}
