% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_functions.R
\name{make_set_percentages_plot}
\alias{make_set_percentages_plot}
\title{Create a plot from a dataset and variable names combining the make_set_percentages and set_percentage_plot functions}
\usage{
make_set_percentages_plot(data, varnames, weight_variable,
  exclude_unique = T)
}
\arguments{
\item{set_percentages}{a names vector with the percentages for each combination}

\item{nsets}{number of sets to look at}

\item{nsets}{number of intersections to look at, the default being 12}

\item{label}{the label to be added to the plot}
}
\value{
A plot object
}
\description{
Create a plot from a dataset and variable names combining the make_set_percentages and set_percentage_plot functions
}
