% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_functions.R
\name{plot_and_save_set_percentages}
\alias{plot_and_save_set_percentages}
\title{Create a plot from a dataset and variable names combining the make_set_percentages and set_percentage_plot functions}
\usage{
plot_and_save_set_percentages(data, varnames, weight_variable = NULL,
  nintersects = 12, exclude_unique = T, label = NULL)
}
\arguments{
\item{data}{a dataframe containing all the 1,0 indicators in varnames}

\item{varnames}{a vector containing the names of variables to be used in the intersection}

\item{weight_variable}{a character string: the name of the variable in the dataset containing the weights, defaults to NULL}

\item{nintersects}{number of intersections to look at, the default being 12}

\item{exclude_unique}{whether the set intersections should include singular sets (i.e. that one variable). Note that if this is set to True, the total set size on the left will be wrong}

\item{label}{the label to be added to the plot}
}
\value{
An UpSetR plot object with the different sets
}
\description{
Create a plot from a dataset and variable names combining the make_set_percentages and set_percentage_plot functions
}
\examples{
see vignette
}
