% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_functions.R
\name{svymean_intersected_sets}
\alias{svymean_intersected_sets}
\title{Create a vector containing the weighted percentages in each set}
\usage{
svymean_intersected_sets(data, intersected_names, weight_variable = NULL,
  weighting_function = NULL)
}
\arguments{
\item{data}{a dataframe containing all the intersected sets}

\item{intersected_names}{the names of the intersected sets: as combinations of variable names combined with '&'}

\item{weight_variable}{character string: the name of the variable in the dataset containing the weights}
}
\value{
A plot of the intersection
}
\description{
Create a vector containing the weighted percentages in each set
}
