% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_functions.R
\name{set_intersection_plot}
\alias{set_intersection_plot}
\title{Create a plot from the percentages in each set}
\usage{
set_intersection_plot(set_percentages, nintersects = 12, nsets = NULL,
  label = NULL)
}
\arguments{
\item{set_percentages}{a names vector with the percentages for each combination}

\item{nintersects}{number of intersections to look at, the default being 12}

\item{nsets}{number of unique sets making up the intersection}

\item{label}{the label to be added to the plot}
}
\value{
A plot object
}
\description{
Create a plot from the percentages in each set
}
